docs: make autoenum a private module, trim comments/docstrings
Rename pyvista/ext/autoenum.py to _autoenum.py: it's an internal
implementation detail of the docs build, not public API.

Also cut the docstrings and comments down to what's actually needed to
understand the code, rather than narrating the rationale for every design
choice.
